backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
timm
pandas
Pillow
scikit-learn